+++ src/main/kotlin/lotto/domain/entity/BonusNumber.kt
@@ -1,17 +1,17 @@
 package lotto.domain.entity
 
-import lotto.domain.validator.BonusNumberValidator
+import lotto.domain.validator.BonusLottoNumberValidator
 
 class BonusNumber(
     val number: Int,
     winningNumbers: List<Int>,
-    validator: BonusNumberValidator = bonusNumberValidator
+    validator: BonusLottoNumberValidator = bonusNumberValidator
 ) {
     init {
         validator.validateBonusNumber(number, winningNumbers)
     }
 
     companion object {
-        private val bonusNumberValidator by lazy { BonusNumberValidator() }
+        private val bonusNumberValidator by lazy { BonusLottoNumberValidator() }
     }
 }
