+++ src/main/kotlin/lotto/domain/entity/Lotto.kt
@@ -1,10 +1,10 @@
 package lotto.domain.entity
 
-import lotto.domain.validator.LottoNumberValidator
+import lotto.domain.validator.LottoNumbersValidator
 
 class Lotto(
     private val numbers: List<Int>,
-    validator: LottoNumberValidator = lottoNumberValidator
+    validator: LottoNumbersValidator = lottoNumbersValidator
 ) {
     init {
         validator.validateLottoNumbers(numbers)
@@ -13,6 +13,6 @@ class Lotto(
     fun getNumbers() = numbers.toList()
 
     companion object {
-        private val lottoNumberValidator by lazy { LottoNumberValidator() }
+        private val lottoNumbersValidator by lazy { LottoNumbersValidator() }
     }
 }
